212-Peering-VPCs/cdk-AWS-Cookbook-212/requirements.txt
-e .
attr
attrs
aws-cdk.assets
aws-cdk.aws-cloudwatch
aws-cdk.aws-ec2
aws-cdk.aws-events
aws-cdk.aws-iam
aws-cdk.aws-kms
aws-cdk.aws-logs
aws-cdk.aws-s3
aws-cdk.aws-s3-assets
aws-cdk.aws-ssm
aws-cdk.cloud-assembly-schema
aws-cdk.core
aws-cdk.cx-api
aws-cdk.region-info
boto3
botocore
cattrs
constructs
jmespath
jsii
publication
python-dateutil
s3transfer
six
typing-extensions
urllib3
aws-cdk.aws_iam 
